backend/requirements.txt
flask==3.1.2
flask-cors==6.0.1
python-dotenv==1.2.1
requests==2.32.5
sqlalchemy==2.0.44
flask-sqlalchemy==3.1.1
psycopg2==2.9.11
openai==2.6.1
httpx==0.27.2
flask-migrate==4.0.4
flask-bcrypt
flask-login

